Optimized the main threading loop
Changed the conversion process from uint64_t to
an array by doing a reinterpret_case<> instead
of the bitwise shifting of the bits.  The casting
should have 0 performance and is determined at
compile time.

The main difference is the ordering in the bytes
which I handled in all the other decrpyting
locations.